requirements.txt
# Библиотека для автоматизации действий в браузере (UI-тесты)
selenium==4.15.0
# Библиотека для выполнения HTTP-запросов (API-тесты)
requests==2.31.0
# Фреймворк для написания и запуска тестов
pytest==7.4.3
# Интеграция Pytest с системой отчётов Allure
allure-pytest==2.13.2
# Библиотека для удобной работы с переменными окружения (из .env файла)
python-dotenv==1.0.0
# Для работы с веб-драйверами
webdriver-manager==4.0.1